refactor: layouts 디렉터리에 있는 컴포넌트를 templates 디렉터리로 이동
