+++ src/components/templates/CenterLayout/CenterLayout.type.ts
