+++ src/components/templates/FullLayout/FullLayout.type.ts
