+++ src/layouts/index.ts
@@ -1,5 +1,3 @@
-import CenterLayout from 'layouts/CenterLayout/CenterLayout';
-import FullLayout from 'layouts/FullLayout/FullLayout';
 import MainLayout from 'layouts/MainLayout/MainLayout';
 
-export { CenterLayout, FullLayout, MainLayout };
+export { MainLayout };
